redistribute when adding
