+++ web-common/src/features/canvas/util.ts
@@ -144,3 +144,38 @@ export function getColumnIndex(
     (rowItem) => rowItem.x === item.x && rowItem.y === item.y,
   );
 }
+
+export function redistributeRowColumns(row: { items: V1CanvasItem[] }) {
+  const totalItems = row.items.length;
+  if (totalItems === 0) return;
+
+  // Total available columns
+  const totalColumns = defaults.COLUMN_COUNT; // Should be 12
+
+  // Distribute columns evenly among items
+  const baseWidth = Math.floor(totalColumns / totalItems); // Minimum width per item
+  const extraColumns = totalColumns % totalItems; // Remainder columns to distribute
+
+  // Sort items by x position to maintain visual order
+  const sortedItems = [...row.items].sort((a, b) => (a.x ?? 0) - (b.x ?? 0));
+
+  // Update widths and x positions
+  let currentX = 0;
+  sortedItems.forEach((item, index) => {
+    // Assign base width to each item
+    item.width = baseWidth;
+
+    // Distribute any leftover columns to the first few items
+    if (index < extraColumns) {
+      item.width += 1;
+    }
+
+    // Update the x position
+    item.x = currentX;
+
+    // Increment currentX for the next item
+    currentX += item.width;
+  });
+
+  return sortedItems;
+}
