+++ web-common/src/features/workspaces/CanvasWorkspace.svelte
@@ -31,7 +31,10 @@
   import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
   import { parseDocument } from "yaml";
   import PreviewButton from "../explores/PreviewButton.svelte";
-  import { findNextAvailablePosition } from "@rilldata/web-common/features/canvas/util";
+  import {
+    findNextAvailablePosition,
+    redistributeRowColumns,
+  } from "@rilldata/web-common/features/canvas/util";
 
   export let fileArtifact: FileArtifact;
 
@@ -122,15 +125,21 @@
       y,
     };
 
+    const updatedItems = [...existingItems, newComponent];
+
+    const rowItems = updatedItems.filter((item) => (item.y ?? 0) === y);
+
+    if (rowItems.length > 1) {
+      redistributeRowColumns({ items: rowItems });
+    }
+
     if (!docJson.items) {
-      parsedDocument.set("items", [newComponent]);
+      parsedDocument.set("items", updatedItems);
     } else {
-      parsedDocument.set("items", [...existingItems, newComponent]);
+      parsedDocument.set("items", updatedItems);
     }
 
     const newIndex = existingItems.length;
-    // $canvasStore.setSelectedComponentIndex(newIndex);
-
     updateEditorContent(parsedDocument.toString(), true);
     await updateComponentFile();
     scrollToComponent(newIndex);
